fix: teach dict layout how to evaluate expressions with non-trivial nullability (#3535)
The dict layout currently evaluates expressions on its values and then
constructs a dict array of the result. This is semantically correct for
non-nullable arrays. For nullable arrays, this forces the null semantics
to be the trivial "poisoning" semantics: any expression whose leaves
evaluate to null must also evaluate to null. This is not true for at
least two expressions in our expression language:
- `is_null(...)`
- `pack{k1: v1, ...}`.

This PR introduces a test case for each expression and makes the most
conservative change, simply evaluating the expression on the constructed
dictionary array. If individual expressions can evaluate more cleverly,
they should!

---------

Signed-off-by: Daniel King <dan@spiraldb.com>
Signed-off-by: mwlon <m.w.loncaric@gmail.com>